Text_Encryption/requirements.txt
pycryptodome
rsa
flask